Fix anchor signature verification
