+++ primitives/src/taiko/proposal.rs
@@ -10,7 +10,7 @@ sol! {
 }
 
 pub fn decode_propose_block_call_args(data: &[u8]) -> Result<proposeBlockCall> {
-    let propose_block_call =
-        proposeBlockCall::abi_decode(data, false).context("failed to decode propose block call")?;
+    let propose_block_call = proposeBlockCall::abi_decode(data, false)
+        .with_context(|| "failed to decode propose block call")?;
     Ok(propose_block_call)
 }
